Move integration test to a new class
Creates a new test that checks santiize() behavior if assertions are
disabled.  The test is executed if Maven is called with:

  mvn clean -DenableAssertions=false verify

Simplify the JenkinsRule based integration test thanks to feedback
from @jglick.  No need to use a WebClient to wrap the execution of
the constructor.
